src/main/java/org/example/deserializers/RoleEnumDeserializer.java
package org.example.deserializers;

import com.fasterxml.jackson.core.JacksonException;
import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;
import org.example.domain.Role;

import java.io.IOException;

public class RoleEnumDeserializer extends JsonDeserializer<Role> {
    @Override
    public Role deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException, JacksonException {
        return Role.getByName(jsonParser.readValueAs(String.class));
    }
}
